Modify FiveDayForecast temperatures Change file names
